chore: simplify PackageVersion check
